tests/wpt/web-platform-tests/workers/Worker_cross_origin_security_err.htm

<!DOCTYPE html>
<title> Worker cross-origin URL </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
test(function() {
  assert_throws("SECURITY_ERR", function() {
    new Worker("ftp://example.org/support/WorkerBasic.js");
  });
});
</script>
